updated to mobile friendly coderloader

+++ src/App.js
@@ -1,12 +1,25 @@
+// TO DO:
+// edit githubs/jobs
+// change skills/resp order in job cards
+// apiURL failover?
+// update sandbox
+// add newlines to text parsing
+// maybe make a text parser that handles links & newlines
+// 'hello my name is' spacing
+
 import './App.css';
 import React from 'react'
-import { Icon, Menu, Segment, Sidebar, Sticky } from 'semantic-ui-react'
+import { Icon, Menu, Segment, Sidebar, Sticky, Confirm, Button} from 'semantic-ui-react'
+
+import NavLinks from './components/NavLinks'
 import Content from './components/Content'
 import Login from './components/Login'
 import LoggedIn from './components/LoggedIn'
 import Editor from './components/Editor'
 
-const apiURL = 'http://portfoliodb.256hz.com/api/v1/'
+// const apiURL = 'http://localhost:3000/api/v1/'
+// const apiURL = 'http://pgdb.256hz.com/api/v1/'
+
 const HEADERS_AUTH = {
   'Authorization': 'Bearer ' + localStorage.jwt,
   'Content-Type': 'application/json'
@@ -22,49 +35,60 @@ const DEFAULT_STATE = {
   links: [],
   users: [],
 
-  message: '',
+  apiURL: 'http://pgdb.256hz.com/api/v1/',
+  confirmOpen: false,
+  contentToDelete: {},
+  creating: {},
+  creatingType: '',
   currentUser: {},
-  sidebarVisible: false,
-  loggedIn: false,
   editorDisabled: true,
   editing: {},
   editingType: '',
-  creating: {},
-  creatingType: ''
+  loggedIn: false,
+  message: '',
+  sidebarVisible: false,
 }
 
-let keys = Object.keys(DEFAULT_STATE)
-let anchors = keys.slice(0, 7)
 // used to automate fetch -- the first 7 entries in default state
-// are the names of the resources we want to fetch.
+// are the names of the resources we get from the API.
+const anchors = Object.keys(DEFAULT_STATE).slice(0, 7)
 
 class App extends React.Component {
     constructor() {
         super()
         this.state = DEFAULT_STATE
+
+        // Try local API first, then fail to the remote 
+        // fetch(this.state.apiURL+'users')
+        // .catch( _ => {
+        //   this.setState({apiURL: 'http://pgdb.256hz.com/api/v1/'}) 
+        //   this.forceUpdate()  
+        // }) 
     }
 
     componentDidMount() {
-      //check for logged in user
+      // check for logged in user
       if (!!localStorage.jwt && !!localStorage.username) {
         this.setState({loggedIn: true, username: localStorage.username})
       } else {
         this.setState({loggedIn: false})
       }
-      //automated fetch
+      // automated fetch
       anchors.forEach( a => {
-        fetch( apiURL + a ).then( res => res.json() )
-        .then( json => this.setState({[a]: json}))
+        fetch( this.state.apiURL + a )
+        .then( res => res.json() )
+        .then( json => this.setState({[a]: json}) )
       })
-      //special fetch for users
-      fetch( apiURL + 'users').then( res => res.json() )
+      // special fetch for users - sets current user as first returned
+      fetch( this.state.apiURL + 'users').then( res => res.json() )
       .then( users => {
         this.setState({users})
         this.setState({currentUser: users[0]})
       })
     }
 
-    openSidebar = () => {
+    toggleSidebar = () => {
+      // open or close sidebar & clear editingType
       this.setState({
         sidebarVisible: !this.state.sidebarVisible,
         editingType: ''
@@ -74,22 +98,32 @@ class App extends React.Component {
     login = (ev, username, password) => {
       ev.preventDefault()
       this.setState({message: ''})
-      fetch(apiURL + 'login', {
+      // no-auth POST to retrieve JWT from Rails
+      fetch(this.state.apiURL + 'login', {
         method: 'POST',
         headers: HEADERS_NOAUTH,
         body: JSON.stringify({ user: {username, password}})
       }).then( res => res.json() )
         .then( json => {
           if (json && json.jwt) {
+            // Login successful
             localStorage.setItem('jwt', json.jwt)
             localStorage.setItem('username', username)
-            this.setState({username: username, loggedIn: true})
+            this.setState({
+              username: username, 
+              loggedIn: true, 
+              sidebarVisible: false,
+              message: ''
+            })
           } else {
+            // Login failed
             localStorage.removeItem('jwt')
             localStorage.removeItem('username')
-            this.setState({username: '', message: json.message, loggedIn: false})
+            this.setState({
+              username: '', 
+              message: json.message, 
+              loggedIn: false})
           }
-          this.setState({sidebarVisible: false})
         })
   }
 
@@ -103,19 +137,21 @@ class App extends React.Component {
     }
 
     startEdit = (content, type) => {
+      // Handles all clicks of the 'edit' button in SectionHeadings
       if (localStorage.getItem('jwt') !== '') {
         this.setState({
           editing: content,
           editingType: type,
           creatingType: '',
           sidebarVisible: true
-        }) //, ()=>console.log('set up edit', this.state.editingType))
+        })
       } else {
         alert('Please log in to edit')
       }
     }
 
     startNew = (type) => {
+      // Handles all clicks of the 'add' button in SectionHeadings
       if (localStorage.getItem('jwt') !== '') {
         this.setState({
           editing: {},
@@ -127,15 +163,16 @@ class App extends React.Component {
           },
           creatingType: type,
           sidebarVisible: true
-        }) //, ()=>console.log('set up new', this.state.creatingType))
+        })
       } else {
         alert('Please log in to add ' + this.state.creatingType)
       }
     }
 
     handleSubmit = (content) => {
-      fetch(apiURL+this.state.editingType+'/'+content.id, {
-        method: "PATCH",
+      // Handles submission of edited content
+      fetch(this.state.apiURL+this.state.editingType+'/'+content.id, {
+        method: 'PATCH',
         headers: HEADERS_AUTH,
         body: JSON.stringify({...content})
       })
@@ -144,19 +181,19 @@ class App extends React.Component {
       .then(json => {
         let editingTypeCopy=this.state.editingType
         switch(editingTypeCopy) {
-          case "users": this.setState({users: [json], currentUser: json})
+          case 'users': this.setState({users: [json], currentUser: json})
             break
-          case "skills": let skillsCopy = this.state.skills.map(skill => {
+          case 'skills': let skillsCopy = this.state.skills.map(skill => {
               return (skill.id === content.id) ? content : skill
             })
             this.setState({skills: skillsCopy})
             break
-          case "jobs": let jobsCopy = this.state.jobs.map(job => {
+          case 'jobs': let jobsCopy = this.state.jobs.map(job => {
               return (job.id === content.id) ? content : job
             })
             this.setState({jobs: jobsCopy})
             break
-          case "githubs": let githubsCopy = this.state.githubs.map(github => {
+          case 'githubs': let githubsCopy = this.state.githubs.map(github => {
               return (github.id === content.id) ? content : github
             })
             this.setState({githubs: githubsCopy})
@@ -168,15 +205,15 @@ class App extends React.Component {
     }
 
     shiftOrder = (incomingGroup, item, next) => {
+      // Change order of skills, jobs, and githubs.
+      // Params are: (group being edited, item, and boolean: whether it's
+      // being shifted up or down in order).
       let group = this.state[incomingGroup].sort( (a,b) => a.order_id - b.order_id )
       let orderIds = group.map( s => s.order_id )
       let curIndex = orderIds.indexOf( item.order_id )
       let maxPos = orderIds.length-1
-      let move = next ? 1 : -1 //if next is true, shift up; else shift down
-
-      // console.log('shifting:', {incomingGroup, item, next})
-      // console.log('to order_id:', item.order_id + move)
-      // console.log({group, orderIds})
+      //if next is true, shift up; else shift down
+      let move = next ? 1 : -1 
 
       if (curIndex === maxPos && next) {
         let t = orderIds[maxPos]
@@ -193,15 +230,15 @@ class App extends React.Component {
       }
 
       group.forEach( (item, index) => {
+        // check if item's order has been changed, and if so, PATCH its order_id
         if (item.order_id !== orderIds[index]) {
-          // console.log(item.name + " changed, fetching")
           item.order_id = orderIds[index]
-          fetch(apiURL + '/' + incomingGroup + '/'+ item.id, {
-            method: "PATCH",
+          fetch(this.state.apiURL + '/' + incomingGroup + '/'+ item.id, {
+            method: 'PATCH',
             headers: HEADERS_AUTH,
             body: JSON.stringify({...item})
           }).then( res => res.json() )
-            .then( console.log )
+          //  .then( console.log )
         }
       })
       this.setState({ [group]: group })
@@ -209,8 +246,8 @@ class App extends React.Component {
 
     handleCreate = (content) => {
       content['order_id']=this.state[this.state.creatingType].length
-      fetch(apiURL+this.state.creatingType, {
-        method: "POST",
+      fetch(this.state.apiURL+this.state.creatingType, {
+        method: 'POST',
         headers: HEADERS_AUTH,
         body: JSON.stringify({
           ...content,
@@ -223,14 +260,22 @@ class App extends React.Component {
         let creatingTypeCopy=this.state.creatingType
         this.setState({
           [creatingTypeCopy]: [...this.state[creatingTypeCopy], json],
-          creatingType: ''
+          creatingType: '',
+          sidebarVisible: false,
         })
       })
     }
 
     handleDelete = (content) => {
-      fetch(apiURL+this.state.editingType+'/'+content.id, {
-        method: "DELETE",
+      this.setState({
+        confirmOpen: true,
+        contentToDelete: content,
+      })
+    }
+
+    confirmDelete = (content) => {
+      fetch(this.state.apiURL+this.state.editingType+'/'+content.id, {
+        method: 'DELETE',
         headers: HEADERS_AUTH
       })
       .then(res => res.json())
@@ -239,67 +284,112 @@ class App extends React.Component {
         copy.splice(copy.findIndex(el => el.id === json.id),1)
         this.setState({
           [this.state.editingType]: copy,
+          confirmOpen: false,
+          contentToDelete: {},
           editingType: '',
-          sidebarVisible: false
+          sidebarVisible: false,
         })
       })
     }
 
     render() {
-        return(
-          <Sidebar.Pushable as={Segment} className="fix-sidebar">
-            <Sticky >
-              <Sidebar as={Menu} animation='overlay'
-                 direction='right' icon='labeled'
-                 inverted vertical
-                 visible={this.state.sidebarVisible}
-                 width='wide'
-               >
-                 <Menu.Item as='a' onClick={this.openSidebar}>
-                   <Icon name='bars' size="mini"/>
-                   Close
-                 </Menu.Item>
+        /*
+          All content is nested within the Sidebar object.  Inside the sidebar (<Sticky>), 
+          components are as follows:
 
-                 <Menu.Item as='a'>
-                    {(this.state.loggedIn && localStorage.getItem('jwt'))
-                      ? <LoggedIn user={this.state.currentUser} logOut={this.logOut}/>
+          - Close button (top)
+          - Login/LoggedIn (login bar/welcome message & logout)
+
+          Then, a ternary shows:
+          - Navigation (if we're not editing or creating anything) or
+          - Editor (if we are editing/creating things) 
+
+          - Close button (bottom)
+          Site content is rendered next.
+          - NamePicIntro
+          - AboutMe
+          - Repos
+          - Jobs
+          - Contact
+          Finally, a confirmation window for deleting a resource.
+        */
+        return(
+          <Sidebar.Pushable as={Segment} className='fix-sidebar'>
+            <Sticky>
+              <Sidebar 
+                animation='overlay'
+                as={Menu} 
+                direction='right' 
+                icon='labeled'
+                inverted 
+                vertical
+                visible={this.state.sidebarVisible}
+                width='wide'
+                >
+                <Menu.Item as='a' onClick={this.toggleSidebar}>
+                  <Icon name='bars' size='mini'/>
+                  Close
+                </Menu.Item>
+
+                <Menu.Item as='a'>
+                    {this.state.loggedIn && localStorage.getItem('jwt')
+                      ? <LoggedIn username={this.state.currentUser.first_name} logOut={this.logOut}/>
                       : <Login login={this.login} message={this.state.message}/>
                     }
-                 </Menu.Item>
-
-                 <Editor
-                  editorDisabled={this.state.editorDisabled}
-                  editing={this.state.editing}
-                  creatingType={this.state.creatingType}
-                  creating={this.state.creating}
-                  handleSubmit={this.handleSubmit}
-                  handleCreate={this.handleCreate}
-                  editingType={this.state.editingType}
-                  startEdit={this.startEdit}
-                  handleDelete={this.handleDelete}
-                 />
-
-               </Sidebar>
-             </Sticky>
-             <Sidebar.Pusher dimmed={false}>
+                </Menu.Item>
+                
+                {this.state.editingType === '' && this.state.creatingType === ''
+                  ? <NavLinks toggleSidebar={this.toggleSidebar}/>
+                  : <Menu.Item>
+                      <Editor
+                        creating=       {this.state.creating}
+                        creatingType=   {this.state.creatingType}
+                        editing=        {this.state.editing}
+                        editorDisabled= {this.state.editorDisabled}
+                        editingType=    {this.state.editingType}
+                        handleCreate=   {this.handleCreate}
+                        handleDelete=   {this.handleDelete}
+                        handleSubmit=   {this.handleSubmit}
+                        shiftOrder=     {this.shiftOrder}
+                        startEdit=      {this.startEdit}
+                      />
+                    </Menu.Item>
+                  }
+                
+                <Menu.Item as='a' onClick={this.toggleSidebar}>
+                  <Icon name='bars' size='mini'/>
+                  Close
+                </Menu.Item>
+              </Sidebar>
+            </Sticky>
+            <Sidebar.Pusher dimmed={false}>
               <Segment basic className={this.state.currentUser.color_theme}>
 
                 <Content
-                  openSidebar={this.openSidebar}
-                  startEdit={this.startEdit}
-                  shiftOrder={this.shiftOrder}
-                  startNew={this.startNew}
-                  jobs={this.state.jobs}
-                  githubs={this.state.githubs}
-                  interests={this.state.interests}
-                  skills={this.state.skills}
-                  honors={this.state.honors}
-                  links={this.state.links}
-                  users={this.state.users}
-                  currentUser= {this.state.currentUser}
-                  editing={this.state.editing}
-                  loggedIn={this.state.loggedIn}
+                  currentUser=  {this.state.currentUser}
+                  editing=      {this.state.editing}
+                  githubs=      {this.state.githubs}
+                  honors=       {this.state.honors}
+                  interests=    {this.state.interests}
+                  jobs=         {this.state.jobs}
+                  links=        {this.state.links}
+                  loggedIn=     {this.state.loggedIn}
+                  skills=       {this.state.skills}
+                  shiftOrder=   {this.shiftOrder}
+                  startEdit=    {this.startEdit}
+                  startNew=     {this.startNew}
+                  toggleSidebar={this.toggleSidebar}
+                  users=        {this.state.users}
                 />
+
+                <Confirm 
+                  cancelButton= {<Button>Go Back</Button>}
+                  confirmButton={<Button negative>Delete</Button>}
+                  onCancel=     {_ => this.setState({confirmOpen: false})}
+                  onConfirm=    {_ => this.confirmDelete(this.state.contentToDelete)}
+                  open=         {this.state.confirmOpen}
+                  size=         'mini'
+                  />                  
 
               </Segment>
             </Sidebar.Pusher>
